feat(scss): first commit regarding SASS addition
Migration of current state of project use of CSS to SASS. This has been
a necessary change in the project structure, and has already been
started to be applied in latest widgets. This first (official) commit
marks the complete migration to SASS. This change gives rise to better
CSS  (compiled/generated) declaration and more logical control over the
projects CSS styles and structrue.