refactor(modal-misc): restructure logic
[1]
Restructure of modals to be more clean and standarized. Includes an
addition of a ModalBackdrop, used by all platform modals. This modal has
been equipped with default platform blocking of scroll, using Svelte
Lifecycle state. In addition, renaming use of CNAME was made.
[2]
Simplification of modal trigger state, which was a single unique
variable for each modal stored in the session. This has been changed to
allow for only a single instance of unique 'string' key, of a target
modal that is to be displayed at a time.
[3]
Includes adjustment to AdminWidgetControl Panel for some modals/pop-ups that require it.