fix(debug): disable use of debug log persistance (+misc)
[1] -----
Remove logic for persisting to 'debug.log' for easier debugging, only temporarily.
[2] -----
Disable 'dev branch reset' workflow, as it interferes with development workflow. Will be updated to reset a 'release' branch instead.
[3] -----
Update '.gitignore' to remove 'debug.log' file from GIT and ran a 'git rm --cached' on it to remove it.