ci: reset branch (dev) reset
[1] -----
Update Github actions workflow to make use of the 'dev' branch reset on 'main' (PR). Initially disabled for the idea of using a 'release' branch to be reset, instead of 'dev', as multiple developers can have pushed to 'dev' when a 'PR' goes to 'main', resetting version control state and unding work.