hotfix(league): date dropdown correct
[1] ────
ISSUE: Wrong order in league season was being displayed, for cases in which 'full-year' is given as a data point. CAUSE: Sportmonks data is inconsistent, and using the 'order' by 'season-id' and/or other data points is not the correct approach. FIX: datecting is it's a full-year, and applying a ordering correction.