refactor(utils): update
[1] ────
Update 'utils/navigation.ts' for updated method 'selectLanguage(..)' for removal of redundnat logic that is never hit after the 'switch(..)' conditions, as all routes are handled by the 'switch' identifier. Also, cleaning up code of the 'serverLang' assignment in some of the 'routes', as these are handled automatically by a 'side-effect' logic trigger.