chore(batch): remove sentry
[1] ────
Remove use of 'sentry' in the project. REASON: (1) Believed to be causing unecessary bloat in the final output bundle; (2) Does not compile correctly at times when heavily using 'chunking' modules in 'vite.config.ts'; (3) Does not work as expected, as errors are dismissed when sentry quota is reached; (4) A custom solution is preferred & will be deployed in near future.