feat(docker,nginx,makefile): update
[1] ────
Update 'nginx.conf' (production) to be of latest configuration that was experimented with 'staging' enviornment. Simplified to just using a single 'server' block, instead of many different server blocks, when all domains are redirected to a single one.
[2] ────
Update 'docker-compose.yml' for addition of 'staging' changes to 'production'.
[3] ────
Update 'Makefile' target for latest 'docker-compose-up' logic flow.